requirements.txt
aiohttp
beautifulsoup4
python-dateutil

git+https://github.com/andymccurdy/redis-py.git
hiredis
git+https://github.com/jsommers/pytricia.git
